Rename references to format to formats
