+++ pkg/formats/formats.go
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: Copyright 2023 The OneSBOM Authors
 // SPDX-License-Identifier: Apache-2.0
 
-package format
+package formats
 
 import "strings"
