+++ pkg/writer/options/options.go
@@ -1,9 +1,11 @@
 package options
 
-import "github.com/bom-squad/protobom/pkg/format"
+import (
+	"github.com/bom-squad/protobom/pkg/formats"
+)
 
 type Options struct {
-	Format format.Format
+	Format formats.Format
 	Indent int
 }
